+++ beaker-gantry/gantry/util.py
@@ -2,8 +2,9 @@ import platform
 import tempfile
 import time
 from pathlib import Path
-from typing import TYPE_CHECKING, Iterable, List, Optional, Tuple
+from typing import TYPE_CHECKING, Iterable, List, Optional, Tuple, Union, cast
 
+import requests
 import rich
 from beaker import (
     Beaker,
@@ -12,6 +13,7 @@ from beaker import (
     DatasetNotFound,
     Digest,
     ExperimentSpec,
+    Priority,
     SecretNotFound,
     TaskResources,
     TaskSpec,
@@ -20,11 +22,11 @@ from beaker import (
 from rich import print, prompt
 from rich.console import Console
 
-from ..exceptions import *
-from ..version import VERSION
 from . import constants
 from .aliases import PathOrStr
 from .constants import GITHUB_TOKEN_SECRET
+from .exceptions import *
+from .version import VERSION
 
 if TYPE_CHECKING:
     from datetime import timedelta
@@ -35,7 +37,7 @@ def unique_name() -> str:
 
     import petname
 
-    return petname.generate() + "-" + str(uuid.uuid4())[:7]
+    return cast(str, petname.generate()) + "-" + str(uuid.uuid4())[:7]
 
 
 def stderr_console() -> Console:
@@ -84,7 +86,7 @@ def display_logs(logs: Iterable[bytes], ignore_timestamp: Optional[str] = None) 
                 return
         except ValueError:
             pass
-        console.print(line, highlight=False)
+        console.print(line, highlight=False, markup=False)
 
     line_buffer = ""
     for bytes_chunk in logs:
@@ -103,15 +105,19 @@ def display_logs(logs: Iterable[bytes], ignore_timestamp: Optional[str] = None) 
     return latest_timestamp
 
 
-def ensure_repo(allow_dirty: bool = False) -> Tuple[str, str, str]:
-    from git import Repo
+def ensure_repo(allow_dirty: bool = False) -> Tuple[str, str, str, bool]:
+    from git.repo import Repo
 
     repo = Repo(".")
     if repo.is_dirty() and not allow_dirty:
         raise DirtyRepoError("You have uncommitted changes! Use --allow-dirty to force.")
     git_ref = str(repo.commit())
     account, repo = parse_git_remote_url(repo.remote().url)
-    return account, repo, git_ref
+    response = requests.get(f"https://github.com/{account}/{repo}")
+    if response.status_code not in {200, 404}:
+        response.raise_for_status()
+    is_public = response.status_code == 200
+    return account, repo, git_ref, is_public
 
 
 def ensure_entrypoint_dataset(beaker: Beaker) -> Dataset:
@@ -167,7 +173,7 @@ def ensure_entrypoint_dataset(beaker: Beaker) -> Dataset:
     ds_files = list(beaker.dataset.ls(gantry_entrypoint_dataset))
     if len(ds_files) != 1:
         raise EntrypointChecksumError(err_msg)
-    if ds_files[0].digest != Digest(sha256_hash.digest()):
+    if ds_files[0].digest != Digest.from_decoded(sha256_hash.digest(), "SHA256"):
         raise EntrypointChecksumError(err_msg)
 
     return gantry_entrypoint_dataset
@@ -186,27 +192,6 @@ def ensure_github_token_secret(
             f"Then upload your token as a Beaker secret using the Beaker CLI or Python client."
         )
     return secret_name
-
-
-def ensure_cluster(beaker: Beaker, task_resources: TaskResources, *clusters: str) -> str:
-    cluster_to_use: str
-    if not clusters:
-        raise ConfigurationError("At least one cluster is required")
-    elif len(clusters) == 1:
-        cluster_to_use = clusters[0]
-    else:
-        available_clusters = sorted(
-            beaker.cluster.filter_available(task_resources, *clusters), key=lambda x: x.queued_jobs
-        )
-        if available_clusters:
-            cluster_to_use = available_clusters[0].cluster.full_name
-            print(f"Using cluster '{cluster_to_use}'")
-        else:
-            cluster_to_use = clusters[0]
-            print_stderr(
-                f"No clusters currently have enough free resources available. Will use '{cluster_to_use}' anyway."
-            )
-    return cluster_to_use
 
 
 def format_timedelta(td: "timedelta") -> str:
@@ -249,7 +234,7 @@ def ensure_datasets(beaker: Beaker, *datasets: str) -> List[Tuple[str, str]]:
 
 def build_experiment_spec(
     task_name: str,
-    cluster_to_use: str,
+    clusters: List[str],
     task_resources: TaskResources,
     arguments: List[str],
     entrypoint_dataset: str,
@@ -265,31 +250,60 @@ def build_experiment_spec(
     venv: Optional[str] = None,
     nfs: Optional[bool] = None,
     datasets: Optional[List[Tuple[str, str]]] = None,
+    env: Optional[List[Tuple[str, str]]] = None,
+    env_secrets: Optional[List[Tuple[str, str]]] = None,
+    priority: Optional[Union[str, Priority]] = None,
+    install: Optional[str] = None,
+    replicas: Optional[int] = None,
+    leader_selection: bool = False,
+    host_networking: bool = False,
 ):
     task_spec = (
         TaskSpec.new(
             task_name,
-            cluster_to_use,
             beaker_image=beaker_image,
             docker_image=docker_image,
             result_path=constants.RESULTS_DIR,
             command=["bash", "/gantry/entrypoint.sh"],
             arguments=arguments,
             resources=task_resources,
+            priority=priority,
+            replicas=replicas,
+            leader_selection=leader_selection,
+            host_networking=host_networking,
         )
+        .with_constraint(cluster=clusters)
         .with_env_var(name="GANTRY_VERSION", value=VERSION)
-        .with_env_var(name="GITHUB_TOKEN", secret=gh_token_secret)
         .with_env_var(name="GITHUB_REPO", value=f"{github_account}/{github_repo}")
         .with_env_var(name="GIT_REF", value=git_ref)
         .with_dataset("/gantry", beaker=entrypoint_dataset)
     )
+
+    if gh_token_secret is not None:
+        task_spec = task_spec.with_env_var(name="GITHUB_TOKEN", secret=gh_token_secret)
+
+    for name, val in env or []:
+        task_spec = task_spec.with_env_var(name=name, value=val)
+
+    for name, secret in env_secrets or []:
+        task_spec = task_spec.with_env_var(name=name, secret=secret)
 
     if conda is not None:
         task_spec = task_spec.with_env_var(
             name="CONDA_ENV_FILE",
             value=str(conda),
         )
-    elif not Path(constants.CONDA_ENV_FILE).is_file():
+    elif Path(constants.CONDA_ENV_FILE).is_file():
+        task_spec = task_spec.with_env_var(
+            name="CONDA_ENV_FILE",
+            value=constants.CONDA_ENV_FILE,
+        )
+    elif Path(constants.CONDA_ENV_FILE_ALTERNATE).is_file():
+        task_spec = task_spec.with_env_var(
+            name="CONDA_ENV_FILE",
+            value=constants.CONDA_ENV_FILE_ALTERNATE,
+        )
+    else:
         task_spec = task_spec.with_env_var(
             name="PYTHON_VERSION", value=".".join(platform.python_version_tuple()[:-1])
         )
@@ -306,9 +320,11 @@ def build_experiment_spec(
             value=venv,
         )
 
-    if nfs is None:
-        if cluster_to_use in constants.NFS_SUPPORTED_CLUSTERS:
-            nfs = True
+    if install is not None:
+        task_spec = task_spec.with_env_var(name="INSTALL_CMD", value=install)
+
+    if nfs is None and clusters and all(["cirrascale" in cluster for cluster in clusters]):
+        nfs = True
 
     if nfs:
         task_spec = task_spec.with_dataset(constants.NFS_MOUNT, host_path=constants.NFS_MOUNT)
@@ -342,7 +358,10 @@ def check_for_upgrades():
 
 
 def ensure_workspace(
-    workspace: Optional[str] = None, yes: bool = False, gh_token_secret: str = GITHUB_TOKEN_SECRET
+    workspace: Optional[str] = None,
+    yes: bool = False,
+    gh_token_secret: str = GITHUB_TOKEN_SECRET,
+    public_repo: bool = False,
 ) -> Beaker:
     beaker = (
         Beaker.from_env(session=True)
@@ -351,7 +370,7 @@ def ensure_workspace(
     )
     try:
         permissions = beaker.workspace.get_permissions()
-        if len(permissions.authorizations) > 1:
+        if not public_repo and len(permissions.authorizations) > 1:
             print_stderr(
                 f"[yellow]Your workspace [b]{beaker.workspace.url()}[/] has multiple contributors! "
                 f"Every contributor can view your GitHub personal access token secret ('{gh_token_secret}').[/]"